docs/book: add grammatical tweaks
* grammatical tweaks

Replace commas with semicolons for readability

* tweaks to opening paragraphs

* benefits paragraph

* capitalisation of machine learning

Treat machine / deep learning as common nouns rather than proper

* avoid dashes

Just so we don't have to use a poor-man's em dash

* rewording intro to layers

Semantics could be different. This needs clarification

* shifting statements about layers

Just putting the general statements about layers at the start of the para

* layerconfig example

Just trying to reword the explanation so it's clearer

* what can layers do para

* layer behaviours para

* activation layers para

* common layers para

Moved to after loss layers, since it refers to them

* utility and container layer paras

* why layers para

* clarification on everything is layers

* tweaks to opening paragraphs

Just migrating my changes to the html over to md. The commit granularity is lost, but that's probably not a really bit deal

* tweaks to layers page

More manual transferring of changes to html pages.

* refresh html files

* why layers para

* what is a layer para

* clarify chapter links

* intro para

* layer config example para

* layer config para

* method paras

* tweak gradients sentence

There may be some semantic differences between this and the original sentence

* tweaks to container layer method desc